+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- Allow marking a character as dead. Dead characters can not spend stars or
+  claim rewards.
 - Show a changelog in the page itself.
 
 ## [1.0.0] - 2019-06-07